% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_log_q.r
\name{compute_log_q}
\alias{compute_log_q}
\title{Compute the logarithm of the variational density}
\usage{
compute_log_q(z, phi)
}
\arguments{
\item{phi}{}
}
\description{
Compute the logarithm of the variational density
}
