% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_compute_log_joint.R
\name{CSS_compute_log_joint}
\alias{CSS_compute_log_joint}
\title{Compute the logarithm of the joint density (CSS)}
\usage{
CSS_compute_log_joint(dist_vec, z, prior_params, B, device)
}
\arguments{
\item{batch_index}{}
}
\description{
Compute the logarithm of the joint density (CSS)
}
