% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_compute_log_q.R
\name{CSS_compute_log_q}
\alias{CSS_compute_log_q}
\title{Compute the logarithm of the variational density (CSS)}
\usage{
CSS_compute_log_q(z, phi)
}
\arguments{
\item{phi}{}
}
\description{
Compute the logarithm of the variational density (CSS)
}
