% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_generate_z.R
\name{CSS_generate_z}
\alias{CSS_generate_z}
\title{Generate samples of z from the variational distribution q (CSS)}
\usage{
CSS_generate_z(phi, S)
}
\arguments{
\item{S}{}
}
\description{
Generate samples of z from the variational distribution q (CSS)
}
