% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_get_phi.R
\name{CSS_get_phi}
\alias{CSS_get_phi}
\title{Get variational parameters from theta. (CSS)}
\usage{
CSS_get_phi(theta)
}
\arguments{
\item{theta}{}
}
\description{
Get variational parameters from theta. (CSS)
}
