% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_init_theta.R
\name{CSS_init_theta}
\alias{CSS_init_theta}
\title{Initialize variational parameters (CSS)}
\usage{
CSS_init_theta(n, p, device)
}
\arguments{
\item{device}{}
}
\description{
Initialize variational parameters (CSS)
}
