% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_idx_to_pairs.r
\name{dist_idx_to_pairs}
\alias{dist_idx_to_pairs}
\title{Convert distance indices to pair indices}
\usage{
dist_idx_to_pairs(idx, n)
}
\arguments{
\item{n}{}
}
\description{
Convert distance indices to pair indices
}
