% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_z.r
\name{generate_z}
\alias{generate_z}
\title{Generate samples of z from the variational distribution q}
\usage{
generate_z(phi, S)
}
\arguments{
\item{S}{}
}
\description{
Generate samples of z from the variational distribution q
}
