% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_log_binconcrete.R
\name{torch_log_binconcrete}
\alias{torch_log_binconcrete}
\title{Logarithm of Binary Concrete Density}
\usage{
torch_log_binconcrete(x, alpha, temperature)
}
\arguments{
\item{temperature}{}
}
\description{
Logarithm of Binary Concrete Density
}
