% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_log_halfcauchy.r
\name{torch_log_halfcauchy}
\alias{torch_log_halfcauchy}
\title{Logarithm of Half-Cauchy Density}
\usage{
torch_log_halfcauchy(x, scale)
}
\arguments{
\item{scale}{}
}
\description{
Logarithm of Half-Cauchy Density
}
