% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_log_invgamma.R
\name{torch_log_invgamma}
\alias{torch_log_invgamma}
\title{Logarithm of Inverse-Gamma Density}
\usage{
torch_log_invgamma(x, a, b)
}
\arguments{
\item{b}{}
}
\description{
Logarithm of Inverse-Gamma Density
}
