% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_log_truncnorm.r
\name{torch_log_truncnorm}
\alias{torch_log_truncnorm}
\title{Logarithm of Truncated Normal Density}
\usage{
torch_log_truncnorm(x, mu, sigma)
}
\arguments{
\item{sigma}{}
}
\description{
Logarithm of Truncated Normal Density
}
