% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rbeta.R
\name{torch_rbeta}
\alias{torch_rbeta}
\title{Generate samples from a Beta distribution}
\usage{
torch_rbeta(sample_shape, shape1, shape2)
}
\arguments{
\item{shape2}{}
}
\description{
Generate samples from a Beta distribution
}
