% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rbinconcrete.R
\name{torch_rbinconcrete}
\alias{torch_rbinconcrete}
\title{Generate samples from a Binary Concrete distribution}
\usage{
torch_rbinconcrete(sample_shape, alpha, temperature)
}
\arguments{
\item{temperature}{}
}
\description{
Generate samples from a Binary Concrete distribution
}
