% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rhalfcauchy.r
\name{torch_rhalfcauchy}
\alias{torch_rhalfcauchy}
\title{Generate samples from a Half-Cauchy distribution}
\usage{
torch_rhalfcauchy(sample_shape, scale)
}
\arguments{
\item{device}{}
}
\description{
Generate samples from a Half-Cauchy distribution
}
