% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIBMDS-package.R
\docType{package}
\name{VIBMDS-package}
\alias{VIBMDS}
\alias{VIBMDS-package}
\title{VIBMDS: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SamMorrissette/VIBMDS}
  \item Report bugs at \url{https://github.com/SamMorrissette/VIBMDS/issues}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
