% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_log_joint.r
\name{compute_log_joint}
\alias{compute_log_joint}
\title{Compute the logarithm of the joint density}
\usage{
compute_log_joint(dist_vec, z, prior_params, B, device)
}
\arguments{
\item{batch_index}{}
}
\description{
Compute the logarithm of the joint density
}
